Combine protocol entities into a single file
