--- modin/core/dataframe/base/dataframe/protocol/column.py
@@ -1,271 +1,0 @@
-# Licensed to Modin Development Team under one or more contributor license agreements.
-# See the NOTICE file distributed with this work for additional information regarding
-# copyright ownership.  The Modin Development Team licenses this file to you under the
-# Apache License, Version 2.0 (the "License"); you may not use this file except in
-# compliance with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under
-# the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific language
-# governing permissions and limitations under the License.
-
-"""
-Dataframe exchange protocol implementation.
-
-See more in https://data-apis.org/dataframe-protocol/latest/index.html.
-"""
-
-from abc import ABC, abstractmethod
-from typing import Any, Optional, Tuple, Dict, Iterable
-
-from .utils import DTypeKind, ColumnNullType
-
-
-class ProtocolColumn(ABC):
-    """
-    A column object, with only the methods and properties required by the interchange protocol defined.
-
-    A column can contain one or more chunks. Each chunk can contain up to three
-    buffers - a data buffer, a mask buffer (depending on null representation),
-    and an offsets buffer (if variable-size binary; e.g., variable-length strings).
-
-    TBD: Arrow has a separate "null" dtype, and has no separate mask concept.
-         Instead, it seems to use "children" for both columns with a bit mask,
-         and for nested dtypes. Unclear whether this is elegant or confusing.
-         This design requires checking the null representation explicitly.
-         The Arrow design requires checking:
-         1. the ARROW_FLAG_NULLABLE (for sentinel values)
-         2. if a column has two children, combined with one of those children
-            having a null dtype.
-         Making the mask concept explicit seems useful. One null dtype would
-         not be enough to cover both bit and byte masks, so that would mean
-         even more checking if we did it the Arrow way.
-    TBD: there's also the "chunk" concept here, which is implicit in Arrow as
-         multiple buffers per array (= column here). Semantically it may make
-         sense to have both: chunks were meant for example for lazy evaluation
-         of data which doesn't fit in memory, while multiple buffers per column
-         could also come from doing a selection operation on a single
-         contiguous buffer.
-         Given these concepts, one would expect chunks to be all of the same
-         size (say a 10,000 row dataframe could have 10 chunks of 1,000 rows),
-         while multiple buffers could have data-dependent lengths. Not an issue
-         in pandas if one column is backed by a single NumPy array, but in
-         Arrow it seems possible.
-         Are multiple chunks *and* multiple buffers per column necessary for
-         the purposes of this interchange protocol, or must producers either
-         reuse the chunk concept for this or copy the data?
-
-    Notes
-    -----
-    This ProtocolColumn object can only be produced by ``__dataframe__``,
-    so doesn't need its own version or ``__column__`` protocol.
-    """
-
-    @property
-    @abstractmethod
-    def size(self) -> int:
-        """
-        Size of the column, in elements.
-
-        Corresponds to `DataFrame.num_rows()` if column is a single chunk;
-        equal to size of this current chunk otherwise.
-
-        Returns
-        -------
-        int
-            Size of the column, in elements.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def offset(self) -> int:
-        """
-        Get the offset of first element.
-
-        May be > 0 if using chunks; for example for a column
-        with N chunks of equal size M (only the last chunk may be shorter),
-        ``offset = n * M``, ``n = 0 .. N-1``.
-
-        Returns
-        -------
-        int
-            The offset of first element.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def dtype(self) -> Tuple[DTypeKind, int, str, str]:
-        """
-        Dtype description as a tuple ``(kind, bit-width, format string, endianness)``.
-
-        * Kind : DTypeKind
-        * Bit-width : the number of bits as an integer
-        * Format string : data type description format string in Apache Arrow C
-                        Data Interface format.
-        * Endianness : current only native endianness (``=``) is supported
-
-        Returns
-        -------
-        tuple
-            ``(kind, bit-width, format string, endianness)``.
-
-        Notes
-        -----
-        - Kind specifiers are aligned with DLPack where possible
-          (hence the jump to 20, leave enough room for future extension).
-        - Masks must be specified as boolean with either bit width 1 (for bit masks)
-          or 8 (for byte masks).
-        - Dtype width in bits was preferred over bytes
-        - Endianness isn't too useful, but included now in case in the future
-          we need to support non-native endianness
-        - Went with Apache Arrow format strings over NumPy format strings
-          because they're more complete from a dataframe perspective
-        - Format strings are mostly useful for datetime specification, and for categoricals.
-        - For categoricals, the format string describes the type of the categorical
-          in the data buffer. In case of a separate encoding of the categorical
-          (e.g. an integer to string mapping), this can be derived from ``self.describe_categorical``.
-        - Data types not included: complex, Arrow-style null, binary, decimal,
-          and nested (list, struct, map, union) dtypes.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def describe_categorical(self) -> Dict[str, Any]:
-        """
-        If the dtype is categorical, there are two options.
-
-        - There are only values in the data buffer.
-        - There is a separate dictionary-style encoding for categorical values.
-
-        TBD: are there any other in-memory representations that are needed?
-
-        Returns
-        -------
-        dict
-            Content of returned dict:
-            - "is_ordered" : bool, whether the ordering of dictionary indices is
-                             semantically meaningful.
-            - "is_dictionary" : bool, whether a dictionary-style mapping of
-                                categorical values to other objects exists
-            - "mapping" : dict, Python-level only (e.g. ``{int: str}``).
-                          None if not a dictionary-style categorical.
-
-        Raises
-        ------
-        ``RuntimeError`` if the dtype is not categorical.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def describe_null(self) -> Tuple[ColumnNullType, Any]:
-        """
-        Return the missing value (or "null") representation the column dtype uses.
-
-        Return as a tuple ``(kind, value)``.
-        * Kind:
-            - 0 : non-nullable
-            - 1 : NaN/NaT
-            - 2 : sentinel value
-            - 3 : bit mask
-            - 4 : byte mask
-        * Value : if kind is "sentinel value", the actual value. If kind is a bit
-          mask or a byte mask, the value (0 or 1) indicating a missing value. None
-          otherwise.
-
-        Returns
-        -------
-        tuple
-            ``(kind, value)``.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def null_count(self) -> int:
-        """
-        Get number of null elements, if known.
-
-        Returns
-        -------
-        int
-
-        Notes
-        -----
-        Arrow uses -1 to indicate "unknown", but None seems cleaner.
-        """
-        pass
-
-    @property
-    @abstractmethod
-    def metadata(self) -> Dict[str, Any]:
-        """
-        Get the metadata for the column.
-
-        See `DataFrame.metadata` for more details.
-
-        Returns
-        -------
-        dict
-        """
-        pass
-
-    @abstractmethod
-    def num_chunks(self) -> int:
-        """
-        Return the number of chunks the column consists of.
-
-        Returns
-        -------
-        int
-           The number of chunks the column consists of.
-        """
-        pass
-
-    @abstractmethod
-    def get_chunks(self, n_chunks: Optional[int] = None) -> Iterable["ProtocolColumn"]:
-        """
-        Return an iterator yielding the chunks.
-
-        By default ``n_chunks=None``, yields the chunks that the data is stored as by the producer.
-        If given, ``n_chunks`` must be a multiple of ``self.num_chunks()``,
-        meaning the producer must subdivide each chunk before yielding it.
-
-        Parameters
-        ----------
-        n_chunks : int, optional
-            Number of chunks to yield.
-
-        Yields
-        ------
-        DataFrame
-            A ``DataFrame`` object(s).
-        """
-        pass
-
-    @abstractmethod
-    def get_buffers(self) -> Dict[str, Any]:
-        """
-        Return a dictionary containing the underlying buffers.
-
-        Returns
-        -------
-        dict
-            - "data": a two-element tuple whose first element is a buffer
-              containing the data and whose second element is the data buffer's associated dtype.
-            - "validity": a two-element tuple whose first element is a buffer
-              containing mask values indicating missing data and
-              whose second element is the mask value buffer's
-              associated dtype. None if the null representation is not a bit or byte mask.
-            - "offsets": a two-element tuple whose first element is a buffer
-              containing the offset values for variable-size binary data
-              (e.g., variable-length strings) and whose second element is the offsets
-              buffer's associated dtype. None if the data buffer does not have
-              an associated offsets buffer.
-        """
-        pass
